darwinConfigurations/epsilon: sunset
The epsilon host was a Macbook Pro bought after I returned from
Silicon Valley in 2017. The Macbook was recycled at Apple UK
before I left Oxford in 2025. The Macbook was haunted by problems
that took Apple to class action lawsuit; the orange screen tint
(especially at winter in Finland!), the flaky "butterfly" keyboard,
among the flexgate issue in which the screen flex cable bent and
broke. I had also dropped the Macbook once in France while doing
my master thesis -- one of the USBC ports broke. I had fell asleep
on my bed with my laptop on my chest, and the laptop fell when I woke.
The bed in my French student housing was where most of my thesis
was written, it being a 9sqm apartment and start of COVID19.
At the beginning of COVID, I lived a month with Soylent.
The same laptop was was used to write my bachelor thesis in Finland.
That is to say, the 15" laptop took me through two degrees.
It was the worst Macbook I ever owned, and worst that I will likely
ever have owned (bye x86!), but despite the flaws it served me good.